backend/requirements.txt
Django>=5.1,<7.0
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg[binary]>=3.2